Made RichTextEditor more robust, Added New Types for Character Dictionaries,  Made Checking for macos
RichTextEditor is now more robust with clear formatting actually removing heading, indent and alignment. Also removed redundant attributes in RichTextEditor that were already in BasicTextEditor. There is now a check in MainWindow to check if it is macOS running the programming to change the 'Ctrl' key to 'Cmd' for easier use.