Added Narrative Forge Projects
There is now a start menu that dictates to create a project or open an existing project. It is also set up so that it opens the previously opened project. DBConnector and SettingsManager are now created in Main and passed to MainWindow when instantied. This is done because StartMenuWindow needs the Settings and DBConnector to work.