Created Narrative Forge Projects.
* Added Narrative Forge Projects

There is now a start menu that dictates to create a project or open an existing project. It is also set up so that it opens the previously opened project. DBConnector and SettingsManager are now created in Main and passed to MainWindow when instantied. This is done because StartMenuWindow needs the Settings and DBConnector to work.

* Fixed the tools.

data_faker.py, db_inspector.py, and db_reset.y have been updated to reflect the new project structure for for Narrative Forge Application.

* Update TODO and add tree.ps1 tool to display tree.

* Update TODO, SettingsManager, Move assets\ to resources\

Update TODO, SettingsManager, and move top level assets\ to resources\ as to not confuse with projects assets.