Added Customs Icons for the MainMenuBar, RichTextEditor, and OutlineManagers (#20)
Custom Icons were added to the MainMenuBar, RichTextEditor, and OutlineManagers.
These icons are stored in resources/icons.

ChapterOutlineManager was reworked to now use a QFrame ands it own Custom ChapterTreeWidget
that allows for chapter reordering on the OutlineManager.

CharacterOutlineManger was reworked to use a QFrame.

LoreOutlineManager was reworked to use a QFrame. A Bug was also fixed when saving the
Parent_Lore_ID of Lore Entries on the OutlineManger.