Relationship Graph Update
A bug was fixed that when a user created new relationship type the 
graph would not allow the user to apply the new type unless they
restarted the application. This has now been fixed.

The Short Label on the RelationshipEdge now will not be upside
down making it much easier to read.

The cursor has been changed after an initial right click to make
it easier to identify when the user is in the create a relationship
part.

The user is now able to right click on on a Relationship to edit it
or otherwise delete the relationship.

There is now a Tool Bar above the Relationship Graph with options
to toggle on/off a background grid, toggle on/off snap to grid, and 
a button to Auto Make a Layout using the C++ Fruchterman-Reingold
algorithm.