UI Updates to Center Titles  and Add Attributes for CharacterEditor
CharacterEditor now shows all of the Character Attributes.

Centering a Lot of the Labels and Tools in ChapterEditor,
LoreEditor, and CharacterEditor.

Small Updating to Graph Logic by Changing Initial Temperature

Fix Bug in the MainMenubar the threw error when trying to create
a new chapter, character, or lore entry. This was done to by creating
a new signal in outline manager to the  create button. This also
switches the view and sets that created item as the current view
item.