feat: Create Notes Table and Note_Tags Table
Created the Notes Table and Note_Tags Table as well as some indexes for it. This 
will allow for the creationof NoteRepository to call Notes from the database using 
this schema in the future. The subsequent documentation has also been updated 
to reflect the Notes Table addition to the database.