Adding Notes View
Create NoteEditor. This editor allows the user to save, create,
update, and delete Note Entities. The user may also tag
notes like with Lore and Chapter.

NoteOutlineManager and NoteEditor were both added to
ViewType and Constants taking the number = 4. They pushed Relationship in EntityType
to = 5 and RELATIONSHIP_GRAPH to 5.

NoteOutlineManager and NoteEditor were both added to MainWindow resulting
so the the logic for loading data to the UI is done.

NoteOutlineManager and NoteEditor were both added to MainMenuBar
so the user has a button to switch to the Note Editing view.

NoteOutlineManager and NoteEditor were added to MainWindow
and connect all of its components. This finaliuzed the implemenation
of Notes.