Fixed Bug in LoreTreeWidget not allowing un-nesting
A bug was found in when trying to unnset items
in LoretreeWidget.DropEvent. The bug was the pyQtSignal
lore_hierarchy_updated had types (int, int) defined but
was trying to emit (int, None) if there was no
parent. This has been fixed to lore_hierarchy_updated
now emits types (int, obj).

Another check was also added in LoretreeWidget.DropEvent
that ensures the Parent id is either and int value or None.