RelationshipEditor Now Uses the Event Bus over Signals
The RelationshipEditor now uses the Event Bus over the signals.
The signals still exist to connect the Nodes and Edges to functions
in RelationshipEditor but Any outside information now uses
the event bus for requesting and receiving information.

AppCoordinator was updated with new functions that
respond to requested information in RelationshipEditor.

Redundant Signals in AppCoordinator, RelationshipEditor,
MainWindow, and ViewManager have been removed.
These signals are no longer needed as the Event Bus now
handles all of this logic.