MainMenuBar now uses the Event Bus
MainMenuBar has been refactored so it now uses the Event Bus
for connections to other components. This required some 
additions to each Editor to ensure the right data is being sent.

Each Button in the MainMenuBar required emitting an event
as opposed to a signal. This clears up confusion of what
signals are supposed to connect to where as we can now
publish and subscribe the events that each component
needs.

The Signal Connections in MainWindow, ViewManager,
AppCoordinator, and MainMenuBar have been removed.
They are no longer needed since we are now using
the event bus.

Event Bus Subscriptions have been added to the OutlineManagers
for creating items. Subscriptions have also been added for